+++ src/helpers/ProximitySearchHelper.php
@@ -14,6 +14,7 @@ namespace doublesecretagency\googlemaps\helpers;
 use Craft;
 use craft\elements\db\ElementQueryInterface;
 use doublesecretagency\googlemaps\fields\AddressField;
+use yii\base\ExitException;
 
 /**
  * Class ProximitySearchHelper
@@ -33,6 +34,11 @@ class ProximitySearchHelper
     private static $_field;
 
     /**
+     * @var array|string Optionally filter results by individual subfields.
+     */
+    private static $_subfieldFilter;
+
+    /**
      * Modify the existing elements query to perform a proximity search.
      *
      * @param $query
@@ -44,6 +50,7 @@ class ProximitySearchHelper
         // Internalize objects
         static::$_query = $query;
         static::$_field = $field;
+        static::$_subfieldFilter = ($options['subfields'] ?? []);
 
         // Join with plugin table
         static::$_query->subQuery->leftJoin(
@@ -58,8 +65,8 @@ class ProximitySearchHelper
         }
 
         // Apply 'subfields' option
-        if (isset($options['subfields'])) {
-            static::_applySubfields($options['subfields']);
+        if (static::$_subfieldFilter) {
+            static::_applySubfields(static::$_subfieldFilter);
         }
 
         // Apply 'requireCoords' option
@@ -234,6 +241,123 @@ class ProximitySearchHelper
     // ========================================================================= //
 
     /**
+     * Perform a geocoding address lookup to determine the coordinates of a given target.
+     *
+     * If necessary, this method will reconfigure the subfields filter
+     * as part of the subfield filter fallback mechanism.
+     *
+     * @param mixed $target
+     * @return array|false Set of coordinates based on specified target.
+     * @throws ExitException
+     */
+    private static function _lookupCoords($target)
+    {
+        // Perform geocoding based on specified target
+        $address = GoogleMaps::lookup($target)->one();
+
+        // Get coordinates of specified target
+        $coords = ($address ? $address->getCoords() : false);
+
+        // If fallback filter is disabled, bail with coordinates
+        if ('fallback' !== static::$_subfieldFilter) {
+            return $coords;
+        }
+
+        // If address contains a valid street address, bail with coordinates
+        if ($address->street1 ?? false) {
+            return $coords;
+        }
+
+        // If no raw address components exist, bail with coordinates
+        if (!($address->raw['address_components'] ?? false)) {
+            return $coords;
+        }
+
+        // Determine type of address result
+        $type = ($address->raw['types'][0] ?? false);
+
+//        Craft::dd($type);
+
+        // List of narrowly focused location types
+        // will be exempt from the fallback filter
+        $focusedTypes = [
+            'premise',      // "123 Main Street"
+            'route',        // "Western Blvd"
+            'intersection', // "Western Blvd and 22nd Street"
+            'locality',     // "Los Angeles"
+            'neighborhood', // "Venice, California"
+        ];
+        /**
+         * We may need to add to this list of focused types.
+         * This is a list of narrowly defined areas, which
+         * can be used for an ACCURATE proximity search.
+         *
+         * Broadly focused types lack the same level of
+         * accuracy, and may be subjected to the subfield
+         * fallback filter mechanism.
+         *
+         * More information:
+         * https://plugins.doublesecretagency.com/google-maps/guides/filter-by-subfields/#subfield-filter-fallback
+         */
+
+        // If the location type is narrowly focused, bail with coordinates
+        if (in_array($type, $focusedTypes)) {
+            return $coords;
+        }
+
+        /**
+         * Still here? It's time to configure the subfields filter.
+         */
+
+        // Restructure the raw address components
+        $address = GeocodingHelper::restructureComponents($address->raw ?? []);
+
+        // Configure subfield filter
+        $filter = [
+            'city'    => $address['city'],
+            'state'   => $address['state'],
+            'zip'     => $address['zip'],
+            'country' => $address['country'],
+        ];
+
+        // Grossly simplify the target string
+        $t = (is_string($target) ? $target : ($target['address'] ?? ''));
+        $t = trim(strtolower($t));
+
+        // Prune unspecified subfields
+        foreach ($filter as $subfield => $value) {
+
+            // If no value was specified
+            if (null === $value) {
+                // Remove from filter
+                unset($filter[$subfield]);
+                // Continue to next
+                continue;
+            }
+
+            // Grossly simplify the filter value
+            $v = trim(strtolower($value));
+
+            // If target and value are identical, filter by THIS PART ONLY!
+            if ($t === $v) {
+                $filter = [$subfield => $value];
+                break;
+            }
+
+        }
+
+        // If subfield filter was properly configured, set it
+        if ($filter) {
+            static::$_subfieldFilter = $filter;
+        }
+
+        // Return coordinates
+        return $coords;
+    }
+
+    // ========================================================================= //
+
+    /**
      * Based on the target provided, determine a center point for the proximity search.
      *
      * @param mixed $target
@@ -247,8 +371,8 @@ class ProximitySearchHelper
             // Target specified as a string
             case 'string':
 
-                // Perform geocoding based on target string, return coordinates
-                return GoogleMaps::lookup($target)->coords();
+                // Return coordinates based on geocoding lookup
+                return static::_lookupCoords($target);
 
             // Target specified as an array
             case 'array':
@@ -257,8 +381,8 @@ class ProximitySearchHelper
                 if (isset($target['lat']) && isset($target['lng'])) {
                     return $target;
                 }
-                // Perform geocoding based on target array, return coordinates
-                return GoogleMaps::lookup($target)->coords();
+                // Return coordinates based on geocoding lookup
+                return static::_lookupCoords($target);
 
         }